lesson 1.2 added new components
