+++ src/AddTodoForm.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const AddTodoForm = () => (
+    <div>
+        <form>
+            <label htmlFor="todoTitle">Title:</label>
+            <input id="todoTitle"></input>
+            <button>Add</button>
+        </form>
+    </div>
+)
+
+
+export default AddTodoForm
